#110: Remove cirecular dependencies and add check to github actions
