+++ .github/workflows/main.yml
@@ -1,0 +1,40 @@
+name: CI/CD
+
+on:
+  push:
+    branches: [main]
+  pull_request:
+    branches: [main]
+
+jobs:
+  build-and-deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout 🛎️
+        uses: actions/checkout@v2.3.1
+
+      - name: Setup ⚙️
+        uses: actions/setup-node@v2
+        with:
+          node-version: '16'
+          cache: 'yarn'
+
+      - name: Install & Test ✅
+        run: |
+          yarn install
+          yarn run test:ci
+
+      - name: Quality Assurance 👌
+        run: |
+          npx dpdm --warning false --tree false --exit-code circular:1 src/**
+          yarn run lint
+
+      - name: Build 🔧
+        run:
+          yarn run build --base=/${{ github.event.repository.name }}/
+
+      - name: Deploy 🚀
+        uses: JamesIves/github-pages-deploy-action@4.1.4
+        with:
+          branch: gh-pages
+          folder: build
