+++ src/timeline/model.tsx
@@ -1,7 +1,4 @@
-import * as React from 'react'
-
-import { ScaleBand, ScaleLinear } from 'd3-scale'
-import { ZoomLevels } from './shared/ZoomScale'
+import React from 'react'
 
 export interface TimelineEvent<EID extends string, LID extends string> {
   eventId: EID
@@ -40,29 +37,3 @@ export type EventComponentFactory<EID extends string, LID extends string, E exte
 
 export type Cursor = 'default' | 'zoom-out' | 'zoom-in' | 'ew-resize' | 'grab'
 export type LaneDisplayMode = 'expanded' | 'collapsed'
-
-export type TimelineLayerType = 'grid' | 'axes' | 'interaction' | 'marks'
-
-export interface CustomLayerProps<EID extends string, LID extends string, E extends TimelineEvent<EID, LID>> {
-  width: number
-  height: number
-  events: ReadonlyArray<E>
-  eventClusters: ReadonlyArray<TimelineEventCluster<LID>>
-  lanes: ReadonlyArray<TimelineLane<LID>>
-  laneDisplayMode?: LaneDisplayMode
-  xScale: ScaleLinear<number, number>
-  yScale: ScaleBand<LID>
-  domain: Domain
-  maxDomain: Domain
-  maxDomainStart: number
-  maxDomainEnd: number
-  currentZoomScale: ZoomLevels
-  nextSmallerZoomScale: ZoomLevels
-  isAnimationInProgress: boolean
-}
-
-export type CustomLayer = <EID extends string, LID extends string, E extends TimelineEvent<EID, LID>>(
-  props: CustomLayerProps<EID, LID, E>
-) => React.ReactNode
-
-export type TimelineLayer = TimelineLayerType | CustomLayer
